+++ core/mode.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Mode(Enum):
+    NORMAL = 0
+    LOW_POWER = 1
+    EMERGENCY = 2
