+++ core/power.py
@@ -1,0 +1,6 @@
+from enum import Enum
+
+class Power(Enum):
+    STARTUP = 8.2
+    NORMAL = 7.6
+    EMERGENCY = 0
